Replace ResponsePlansController w/ DraftsController
