+++ app/views/drafts/_deescalation_technique_fields.html.erb
