+++ app/views/drafts/_image_fields.html.erb
