+++ app/views/drafts/_response_strategy_fields.html.erb
