+++ app/views/drafts/_title.html.erb
@@ -8,7 +8,7 @@
   </div>
 
   <div class="page-title-actions">
-    <%= link_to t("response_plans.draft.edit"), edit_response_plan_path(draft) %>
+    <%= link_to t("response_plans.draft.edit"), edit_draft_path(draft) %>
     or
     <strong>
       <%= link_to(
