+++ app/views/drafts/_trigger_fields.html.erb
